Listing service that can be configured from env vars
