+++ .pylintrc
@@ -56,7 +56,6 @@ confidence=
 # no Warning level messages displayed, use"--disable=all --enable=classes
 # --disable=W"
 disable=invalid-name,
-        import-error,
         print-statement,
         parameter-unpacking,
         unpacking-in-except,
