+++ browse/services/listing/__init__.py
@@ -1,104 +1,21 @@
-"""Serves lists of articles for categories and time periods.
+"""Serves lists of articles for categories and time periods."""
 
-Currently (2018-10) getting everything for a listing from the DB is
-not possible. There is no table that correctly records the publish
-history in the legacy DB.
-
-The legacy listing files are used only for the IDs of the papers
-announced. The rest of the metadata is not kept updated. An example of
-this causing a problem is if an article published on 2018-01-01, then
-crossed on 2018-01-02, then replaced with a differnt title on
-2018-01-03. The cross on 2018-01-02 in the listing file will have the
-old title.
-
-Why month granularity? The legacy listing files have only month
-granularity for when a paper was announced. In the future there might
-be better date granularity for new papers.
-"""
-
-from typing import cast, Optional
-
-from flask import g
-
-from browse.domain.listing import NewResponse, ListingResponse, ListingCountResponse
-
-
-class ListingService:
-    """Class for arXiv document listings."""
-
-    @classmethod
-    def version(cls) -> str:
-        """Version."""
-        return "0.2"
-
-    def list_articles_by_year(self,
-                              archiveOrCategory: str,
-                              year: int,
-                              skip: int,
-                              show: int,
-                              if_modified_since: Optional[str] = None) -> ListingResponse:
-        """Get listing items for a whole year.
-
-        if_modified_since is the if_modified_since header value passed by the web client
-        It should be in RFC 1123 format.
-        """
-        raise NotImplementedError
-
-    def list_articles_by_month(self,
-                               archiveOrCategory: str,
-                               year: int,
-                               month: int,
-                               skip: int,
-                               show: int,
-                               if_modified_since: Optional[str] = None) -> ListingResponse:
-        """Get listings for a month.
-
-        if_modified_since is the if_modified_since header value passed by the web client
-        It should be in RFC 1123 format.
-        """
-        raise NotImplementedError
-
-    def list_new_articles(self,
-                          archiveOrCategory: str,
-                          skip: int,
-                          show: int,
-                          if_modified_since: Optional[str] = None) -> NewResponse:
-        """Gets listings for the most recent announcement/publish.
-
-        if_modified_since is the if_modified_since header value passed by the web client
-        It should be in RFC 1123 format.
-        """
-        raise NotImplementedError
-
-    def list_pastweek_articles(self,
-                               archiveOrCategory: str,
-                               skip: int,
-                               show: int,
-                               if_modified_since: Optional[str] = None) -> ListingResponse:
-        """Gets listings for the 5 most recent announcement/publish.
-
-        if_modified_since is the if_modified_since header value passed by the web client
-        It should be in RFC 1123 format.
-        """
-        raise NotImplementedError
-
-    def monthly_counts(self,
-                       archive: str,
-                       year: int) -> ListingCountResponse:
-        """Gets monthly listing counts for the year."""
-        raise NotImplementedError
-
+from typing import cast, Any
+from browse.services.listing.base_listing import ListingService
+from browse.config import Settings
 
 
 def get_listing_service() -> ListingService:
-    """Get the listing service.
-
-    There is probably a better way to do this.
-    """
+    """Get the listing service configured for the app context"""
+    from browse.config import settings
+    from flask import g
     if 'listing_service' not in g:
-        # importing at runtime to avoid cyclic imports that kill python
-        import importlib
-        fl = importlib.import_module("browse.services.listing.fake_listings")
-        g.listing_service = fl.FakeListingFilesService()  # type: ignore
+        g.listing_service = settings.DOCUMENT_LISTING_SERVICE(settings, g)
 
     return cast(ListingService, g.listing_service)
+
+
+def fake(settings: Settings, _: Any) -> ListingService:
+    """Integration for fake listing service"""
+    from .fake_listings import FakeListingFilesService
+    return FakeListingFilesService()
